// apps/web/src/features/admin/types.ts
export type ModelParameterDataType =
  "string" | "number" | "boolean" | "array<string>" | "array<number>"

export type ModelParameterUIComponent = "select" | "slider" | "uploader"
// ...
export type CreateParameterOptionInput = {
  label: string
  value: string
  description?: string | null
}

export type UpdateParameterOptionInput = Partial<CreateParameterOptionInput>

export type CreateParameterInput = {
  key: string
  label: string
  type: ModelParameterDataType
  description?: string | null
  xUiComponent?: ModelParameterUIComponent | null
  options?: Array<CreateParameterOptionInput>
}